adr: Make  min_dr and max_dr overridable per device
min_dr and max_dr may sometimes require an override on a per-device
basis. This can be stored in the per device tags/variables. This change
makes it accessible to the ADR algorithm in the Request

We also make the min_tx_power_index available next to the
max_tx_power_index, allowing us to clamp them properly.

Signed-off-by: D.S. Ljungmark <spider@modio.se>