WIP: Implement UseExternalJoinServer via tag
This seems awkward in the original logic in the go code, becaise the
default value (False)  means always using getDefaultClient.  The logic
is further confused by the new code no longer having a default client to
fall back to.

Instead, we just shortcut the logic (for now) as if there had not been a
found joinserver for the device with a tag.