Fix the Pulsar integration tests
If using regexp-based matching on topics, the subscriptions will not
catch the topic until the tests have run at least once, causing failures
in cases where the first message has not connected and sent to a topic.

This is sort of expected behaviour, so we connect the test-cases to the
topic explicitly to make sure they exist.

Signed-off-by: D.S. Ljungmark <spider@skuggor.se>